requirements.txt
sanic>=23.12
pymongo>=4.6
python-dotenv>=1.0
certifi>=2024.2.2
aiohttp>=3.9